chore: Deprecate WorkflowLogInterceptor and appendDefaultInterceptors
