+++ packages/test/src/test-local-activities.ts
@@ -14,7 +14,6 @@ import { msToNumber } from '@temporalio/common/lib/time';
 import { temporal } from '@temporalio/proto';
 import { TestWorkflowEnvironment, workflowInterceptorModules } from '@temporalio/testing';
 import {
-  appendDefaultInterceptors,
   bundleWorkflowCode,
   DefaultLogger,
   LogLevel,
@@ -61,9 +60,9 @@ function helpers(t: ExecutionContext<Context>): Helpers {
         connection: t.context.env.nativeConnection,
         workflowBundle: t.context.workflowBundle,
         taskQueue,
-        interceptors: appendDefaultInterceptors({
+        interceptors: {
           activity: interceptors?.activity ?? [ConnectionInjectorInterceptor.createFactory(t.context.env.connection)],
-        }),
+        },
         showStackTraceSources: true,
         ...rest,
       });
