+++ packages/worker/src/index.ts
@@ -37,7 +37,6 @@ export {
   WorkerStatus,
 } from './worker';
 export {
-  appendDefaultInterceptors,
   CompiledWorkerOptions,
   ReplayWorkerOptions,
   WorkerOptions,
@@ -46,17 +45,7 @@ export {
   WorkflowBundlePath,
 } from './worker-options';
 export { ReplayError, ReplayHistoriesIterable, ReplayResult } from './replay';
-export {
-  WorkflowInboundLogInterceptor, // eslint-disable-line deprecation/deprecation
-  WorkflowLogInterceptor,
-  workflowLogAttributes,
-} from './workflow-log-interceptor';
-export {
-  BundleOptions,
-  bundleWorkflowCode,
-  defaultWorkflowInterceptorModules,
-  WorkflowBundleWithSourceMap,
-} from './workflow/bundler';
+export { BundleOptions, bundleWorkflowCode, WorkflowBundleWithSourceMap } from './workflow/bundler';
 
 // Anything below this line is deprecated
 
@@ -77,6 +66,11 @@ export {
 } from './activity';
 export {
   /**
+   * @deprecated Including `appendDefaultInterceptors()` in the worker options is no longer required. To configure a
+   *             custom logger, set the {@see Runtime.logger} property instead.
+   */
+  appendDefaultInterceptors, // eslint-disable-line deprecation/deprecation
+  /**
    * @deprecated Including `defaultSinks()` in the worker options is no longer required. To configure
    *             a custom logger, set the {@link Runtime.logger} property instead.
    */
@@ -93,3 +87,27 @@ export {
    */
   LoggerSinks, // eslint-disable-line deprecation/deprecation
 } from '@temporalio/workflow';
+export {
+  /**
+   * @deprecated `WorkflowInboundLogInterceptor` is deprecated. Workflow lifecycle events are now automatically logged
+   *             by the SDK. To customize workflow log attributes, simply register a custom `WorkflowInterceptors` that
+   *             intercepts the `outbound.getLogAttributes()` method.
+   */
+  WorkflowInboundLogInterceptor, // eslint-disable-line deprecation/deprecation
+  /**
+   * @deprecated `WorkflowLogInterceptor` is deprecated. Workflow lifecycle events are now automatically logged
+   *             by the SDK. To customize workflow log attributes, simply register a custom `WorkflowInterceptors` that
+   *             intercepts the `outbound.getLogAttributes()` method.
+   */
+  WorkflowLogInterceptor, // eslint-disable-line deprecation/deprecation
+} from './workflow-log-interceptor';
+export {
+  /**
+   * @deprecated This function is meant for internal usage. Don't use it.
+   */
+  workflowLogAttributes,
+} from '@temporalio/workflow/lib/logs';
+/**
+ * @deprecated Including `defaultWorkflowInterceptorModules` in BundlerOptions.workflowInterceptorModules is no longer required.
+ */
+export const defaultWorkflowInterceptorModules = [];
