+++ packages/workflow/src/errors.ts
@@ -1,5 +1,6 @@
 import { ActivityFailure, CancelledFailure, ChildWorkflowFailure } from '@temporalio/common';
 import { SymbolBasedInstanceOfError } from '@temporalio/common/lib/type-helpers';
+import { coresdk } from '@temporalio/proto';
 
 /**
  * Base class for all workflow errors
@@ -12,6 +13,16 @@ export class WorkflowError extends Error {}
  */
 @SymbolBasedInstanceOfError('DeterminismViolationError')
 export class DeterminismViolationError extends WorkflowError {}
+
+/**
+ * A class that acts as a marker for this special result type
+ */
+@SymbolBasedInstanceOfError('LocalActivityDoBackoff')
+export class LocalActivityDoBackoff extends Error {
+  constructor(public readonly backoff: coresdk.activity_result.IDoBackoff) {
+    super();
+  }
+}
 
 /**
  * Returns whether provided `err` is caused by cancellation
